refactor: switch to epicsThread + PRIu64/PRId64
Two related cleanups in pvacms cluster code:

1. Convert std::thread to epicsThread (matches the rest of the codebase
   which uses epicsThread for joinable workers):
   * rejoin_watchdog_thread_ now an std::unique_ptr<epicsThread> driven
     by an epicsThreadRunable subclass.  Joined via exitWait() in dtor.
   * rejoin_thread_ + old_rejoin_threads_ likewise hold
     std::unique_ptr<epicsThread> (epicsThread is non-movable but the
     unique_ptr is, so the side-vector pattern still works).
   * rejoin_thread_mutex_ now an epicsMutex (was std::mutex).

2. Standardise integer log-formatting on uint64_t / int64_t:
   * Replace %llu + static_cast<unsigned long long>(x) with
     %" PRIu64 " + the original uint64_t value (cinttypes, C++11).
   * Replace %lld + static_cast<long long>(x) with %" PRId64 " + the
     original int64_t value.
   * PERMANENTLY_VALID_STATUS macro: (~(unsigned long long)0) >> 1
     becomes (~(uint64_t)0) >> 1.
   * One off_t printf cast in pvacms.cpp uses PRIdMAX + intmax_t since
     off_t is platform-dependent in a way uint64_t/int64_t are not.

All 238 harness assertions pass after both changes.

+++ src/authn/ccrmanager.cpp
@@ -6,6 +6,8 @@
 
 #include "ccrmanager.h"
 
+#include <cinttypes>
+
 #include <pvxs/client.h>
 #include <pvxs/log.h>
 #include <pvxs/nt.h>
@@ -66,7 +68,7 @@ std::tuple<time_t, std::string> CCRManager::createCertificate(const std::shared_
         log_info_printf(auth_log, "X.509 certificate RENEWED (%s)\n", value["state"].as<std::string>().c_str());
     }
     log_debug_printf(auth_log, "%s\n", value["value.index"].as<std::string>().c_str());
-    log_debug_printf(auth_log, "%llu\n", (unsigned long long)value["serial"].as<serial_number_t>());
+    log_debug_printf(auth_log, "%" PRIu64 "\n", value["serial"].as<serial_number_t>());
     log_debug_printf(auth_log, "%s\n", value["issuer"].as<std::string>().c_str());
     log_debug_printf(auth_log, "%s\n", value["cert_id"].as<std::string>().c_str());
     log_debug_printf(auth_log, "%s\n", value["status_pv"].as<std::string>().c_str());
